# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fish App Documentation",
        "description": "Documentation",
        "contact": {
            "email": "kanamboma@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Api Server"
        }
    ],
    "paths": {
...
        "/especes": {
            "get": {
                "tags": [
                    "Espèce"
                ],
                "summary": "Retourner la liste des especes",
                "description": "Retourner la liste des especes",
                "operationId": "getAllespeces",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Espèce"
                ],
                "summary": "Ajouter une espece",
                "description": "Ajouter une espece",
                "operationId": "addespeces",
                "requestBody": {
                    "description": "Attribut de la classe espece",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "XXXX"
                                    },
                                    "description": {
                                        "type": "text",
                                        "example": "XXXX"
                                    },
                                    "illustration": {
                                        "type": "string",
                                        "example": "XXXX"
                                    },
                                    "remarque": {
                                        "type": "text",
                                        "example": "XXXX"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": "XXXX"
                                    },
                                    "famille_id": {
                                        "type": "integer",
                                        "example": "XXXX"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/especes/{id}": {
            "get": {
                "tags": [
                    "Espèce"
                ],
                "summary": "Retourner une espece",
                "description": "Retourner une des especes",
                "operationId": "getAllespece",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Espèce"
                ],
                "summary": "Modifier une espece",
                "description": "Modifier une espece",
                "operationId": "updateespece",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Attribut de la classe espece",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "XXXX"
                                    },
                                    "description": {
                                        "type": "text",
                                        "example": "XXXX"
                                    },
                                    "illustration": {
                                        "type": "string",
                                        "example": "XXXX"
                                    },
                                    "remarque": {
                                        "type": "text",
                                        "example": "XXXX"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": "XXXX"
                                    },
                                    "famille_id": {
                                        "type": "integer",
                                        "example": "XXXX"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Espèce"
                ],
                "summary": "Supprimer une espece",
                "description": "Supprimer une des especes",
                "operationId": "deleteespece",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
    }
}